+++ contribs/ev/src/main/java/org/matsim/contrib/ev/dvrp/EvDvrpFleetModule.java
@@ -21,11 +21,11 @@
 package org.matsim.contrib.ev.dvrp;
 
 import org.matsim.api.core.v01.network.Network;
-import org.matsim.contrib.dvrp.data.file.FleetReader;
 import org.matsim.contrib.dvrp.fleet.DvrpVehicleImpl;
 import org.matsim.contrib.dvrp.fleet.DvrpVehicleSpecification;
 import org.matsim.contrib.dvrp.fleet.Fleet;
 import org.matsim.contrib.dvrp.fleet.FleetImpl;
+import org.matsim.contrib.dvrp.fleet.FleetReader;
 import org.matsim.contrib.dvrp.fleet.FleetSpecification;
 import org.matsim.contrib.dvrp.fleet.FleetSpecificationImpl;
 import org.matsim.contrib.dvrp.router.DvrpRoutingNetworkProvider;
